refactor: update icon of rsETH token
