+++ .changeset/orange-dingos-listen.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": patch
+---
+
+Update icon of rsETH token
