+++ apps/evm/src/libs/tokens/infos/commonTokens/ethereum.ts
@@ -4,7 +4,7 @@ import daiLogo from 'libs/tokens/img/dai.svg';
 import ethLogo from 'libs/tokens/img/eth.svg';
 import fraxLogo from 'libs/tokens/img/frax.svg';
 import ptWeethLogo from 'libs/tokens/img/ptWeeth.svg';
-import rsEthLogo from 'libs/tokens/img/rsEth.png';
+import rsEthLogo from 'libs/tokens/img/rsEth.svg';
 import sFraxLogo from 'libs/tokens/img/sFrax.svg';
 import tusdLogo from 'libs/tokens/img/tusd.svg';
 import usdcLogo from 'libs/tokens/img/usdc.svg';
