+++ turbo.json
@@ -1,10 +1,6 @@
 {
   "$schema": "https://turbo.build/schema.json",
-  "globalDependencies": [
-    ".nvmrc",
-    "tsconfig.json",
-    "vite.config.ts"
-  ],
+  "globalDependencies": [".nvmrc", "tsconfig.json", "vite.config.ts"],
   "tasks": {
     "generate": {
       "outputs": [
@@ -16,26 +12,17 @@
       ]
     },
     "start": {
-      "dependsOn": [
-        "^build",
-        "generate"
-      ],
+      "dependsOn": ["^build", "generate"],
       "cache": false,
       "persistent": true
     },
     "test": {
       "cache": false,
-      "outputs": [
-        "coverage/**/*"
-      ]
+      "outputs": ["coverage/**/*"]
     },
     "build": {
-      "dependsOn": [
-        "^build"
-      ],
-      "outputs": [
-        "build/**/*"
-      ]
+      "dependsOn": ["^build"],
+      "outputs": ["build/**/*"]
     },
     "lint": {
       "cache": false
